// packages/privacy-preservation/src/types/userTypes.ts
// Type Imports
import type { ThemeColor } from './types'

export type UsersType = {
  id: number
  role: string
  email: string
  status: string
  avatar: string
  company: string
  country: string
  contact: string
  fullName: string
  username: string
  currentPlan: string
  avatarColor?: ThemeColor
}

export type FilesType = {
  id: number
  type: string
  name: string
  description: string
  lastUpdated: string
}